requirements.txt
altair==4.2.0
astor==0.8.1
attrs==21.2.0
base58==2.1.1
bcrypt==3.2.0
blinker==1.4
cachetools==4.2.4
certifi==2022.9.14
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.3
decorator==5.1.0
entrypoints==0.3
extra-streamlit-components==0.1.56
gitdb==4.0.9
GitPython==3.1.26
idna==3.2
importlib-metadata==4.11.1
Jinja2==3.0.1
jsonschema==3.2.0
markdown-it-py==2.0.1
MarkupSafe==2.0.1
numpy==1.21.3
packaging==21.0
pandas==1.3.4
Pillow==8.4.0
protobuf==3.18.0
pyarrow==3.0.0
pydeck==0.7.1
Pygments==2.10.0
PyJWT==2.3.0
Pympler==1.0.1
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
requests==2.26.0
rich==12.5.1
six==1.16.0
smmap==5.0.0
streamlit==1.5.0
streamlit-authenticator==0.2.1
streamlit-lottie==0.0.3
streamlit-option-menu==0.3.2
toml==0.10.2
toolz==0.11.2
tornado==6.1
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.6
validators==0.18.2
watchdog==2.1.6
zipp==3.7.0
